// src/services/models/workflow-data-model.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface WorkFlowDataModel
 */

/**
    * @export
    * @enum {string}
    */
export enum WorkFlowDataModelTypeEnum {
    APPROVED = 'APPROVED',
    REJECTED = 'REJECTED',
    FAVOURITES = 'FAVOURITES',
    DRAFT = 'DRAFT',
    PENDING = 'PENDING',
    EXPIRED='EXPIRED',
    INFO_REQUEST='INFO_REQUEST'
}

export enum CatalogStatusEnum {
    APPROVED = 0,
    REJECTED =2 ,
    FAVOURITES = 1,
    DRAFT = 5,
    PENDING = 3,
    EXPIRED=6,
    INFO_REQUEST=4
}